replace lua with luajit
Former-commit-id: 672da1c7e4c41a00d0c0ca69e6e1bc76a48c0435 [formerly c1ba7ca49d097a241e14c66fe3990320fb606cb8]
Former-commit-id: c8839590883f4290a3485550433e85cb2e50051d

+++ lib/cocos2d-x/cocos2dx/platform/platform.cpp
@@ -31,7 +31,7 @@ NS_CC_BEGIN
 // from 2dx
 static unsigned int _Hash(const char *key)
 {
-    unsigned int len = strlen(key);
+    size_t len = strlen(key);
     const char *end=key+len;
     unsigned int hash;
